--- src/main/java/entities/archive/ArchiveFactory.java
@@ -1,7 +1,0 @@
-package entities.archive;
-
-public class ArchiveFactory {
-    public Archive create() {
-        return new Archive();
-    }
-}
