--- src/main/java/entities/archive/ArchiveFactoryInterface.java
@@ -1,7 +1,0 @@
-package entities.archive;
-
-import entities.archive.Archive;
-
-public interface ArchiveFactoryInterface {
-    Archive create();
-}
