--- src/main/java/entities/archive/ArchiveInterface.java
@@ -1,15 +1,0 @@
-package entities.archive;
-
-import entities.course.Course;
-
-import java.util.List;
-
-public interface ArchiveInterface {
-    static List<Course> getCourses() {
-        return null;
-    }
-
-    void addCourse(Course course);
-
-    void RemoveCourse(Course course);
-}
