+++ src/main/java/entities/assessment/instanceList/InstanceListFactory.java
@@ -1,4 +1,4 @@
-package entities.instanceList;
+package entities.assessment.instanceList;
 
 public class InstanceListFactory {
     public InstanceList create(String pluralTitle, int totalNumberOfInstances) {
