+++ src/main/java/entities/assessment/instanceList/InstanceListFactoryInterface.java
@@ -1,4 +1,4 @@
-package entities.instanceList;
+package entities.assessment.instanceList;
 
 public interface InstanceListFactoryInterface {
     InstanceList create(String pluralTitle, int totalNumberOfInstances);
