--- src/main/java/entities/semester/SemesterFactory.java
@@ -1,7 +1,0 @@
-package entities.semester;
-
-public class SemesterFactory {
-    public Semester create() {
-        return new Semester();
-    }
-}
