--- src/main/java/entities/semester/SemesterFactoryInterface.java
@@ -1,6 +1,0 @@
-package entities.semester;
-
-public interface SemesterFactoryInterface {
-
-    Semester create();
-}
